implementasi artikel

+++ database/migrations/2024_06_06_074819_create_psikologs_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->string('program_studi');
             $table->integer('tahun_lulus');
             $table->string('tempat_praktik');
-            $table->integer('no_str');
+            $table->bigInteger('no_str');
             $table->string('dokumen_cv');
             $table->string('dokumen_ijazah');
             $table->string('dokumen_str_aktif');
